requirements.txt
environ
Django
django_environ
# psycopg2-binary
Unipath
django-crispy-forms
crispy-tailwind
djangorestframework
djangorestframework-simplejwt
workdays
django-extensions
Werkzeug
django-watson
django-jsoneditor
django-debug-toolbar
django-timedeltafield
django-annoying
xlrd
django-simple-history
ipython
jupyter
django-guardian
django-sendgrid-v5
openpyxl
django-tinymce
crispy-bootstrap5
django-recaptcha==3.0.0
django-anymail[mailgun]
django_compressor
